+++ resources/js/controllers/carousel-controller.js
@@ -1,10 +1,8 @@
 export default class extends window.Controller
 {
     connect() {
-        this.element.querySelectorAll('.oi-carousel').forEach(el => {
-            new window.Bootstrap.Carousel(el, {
-
-            })
+        this.carousel = new window.Bootstrap.Carousel(this.element.querySelector('.oi-carousel'), {
+            // @see https://getbootstrap.com/docs/5.0/components/carousel/#options
         })
     }
 }