+++ resources/views/components/avatar.blade.php
@@ -1,10 +1,9 @@
 @component($typeForm, get_defined_vars())
-    <div class="avatar oi-avatar" style="--oi-avatar-height: {{ $height }}; --oi-avatar-width: {{ $width }};">
-        <img
-            class="border object-fit-cover w-100 h-100"
-            src="{{ $src }}"
-            alt="{{ $alt }}"
-        />
+    <div
+        class="avatar oi-avatar"
+        style="--oi-avatar-height: {{ $height }}; --oi-avatar-width: {{ $width }};"
+    >
+        @include('orchid-images::partials.image', ['fit' => 'object-fit-cover'])
 
         @if ($badge !== false)
             <b class="oi-avatar__badge rounded-pill bg-{{ $badgeType }}">