+++ resources/views/components/gallery.blade.php
@@ -4,19 +4,9 @@
         style="grid-template-columns: {{ $templateColumns }};"
     >
         @forelse ($elements as $item)
-            <li class="oi-image" style="--oi-gallery-aspect-ratio: {{ $aspectRatio }};">
-                <img
-                    @class([
-                        'border d-block rounded w-100 h-100',
-                        $fit,
-                    ])
-                    src="{{ $item->url() }}"
-                    alt="{{ $item->alt }}"
-                    title="{{ $item->title }}"
-                />
-            </li>
+            @include('orchid-images::partials.gallery.gallery-image')
         @empty
-            <li class="oi-gallery__empty">{!! $empty !!}</li>
+            @include('orchid-images::partials.gallery.empty')
         @endforelse
     </ul>
 @endcomponent