+++ resources/views/components/image.blade.php
@@ -1,12 +1,8 @@
 @component($typeForm, get_defined_vars())
-    <div class="oi-image" style="--oi-image-height: {{ $height }}; --oi-image-width: {{ $width }};">
-        <img
-            @class([
-                'border rounded w-100 h-100',
-                $fit,
-            ])
-            src="{{ $src }}"
-            alt="{{ $alt }}"
-        />
+    <div
+        class="oi-image"
+        style="--oi-image-height: {{ $height }}; --oi-image-width: {{ $width }};"
+    >
+        @include('orchid-images::partials.image')
     </div>
 @endcomponent