+++ resources/views/partials/gallery/empty.blade.php
@@ -1,0 +1,3 @@
+<li class="oi-gallery__empty">
+    {!! $empty !!}
+</li>