+++ resources/views/partials/gallery/gallery-image.blade.php
@@ -1,0 +1,6 @@
+<li
+    class="oi-image"
+    style="--oi-gallery-aspect-ratio: {{ $aspectRatio }};"
+>
+    @include('orchid-images::partials.image', ['src' => $item->url(), 'alt' => $item->alt, 'title' => $item->title])
+</li>