+++ resources/views/partials/gallery/lightbox-image.blade.php
@@ -1,0 +1,11 @@
+<li>
+    <a
+        href="{{ $item->url() }}"
+        class="glightbox oi-image d-block"
+        style="--oi-gallery-aspect-ratio: {{ $aspectRatio }};"
+        data-gallery="oi-gallery-{{ $id }}"
+        {{ $dataAttributes }}
+    >
+        @include('orchid-images::partials.image', ['src' => $item->url(), 'alt' => $item->alt, 'title' => $item->title])
+    </a>
+</li>