+++ src/Screen/Components/Carousel.php
@@ -4,92 +4,102 @@ declare(strict_types=1);
 
 namespace Czernika\OrchidImages\Screen\Components;
 
-use Czernika\OrchidImages\Support\Helper;
-use Czernika\OrchidImages\Support\Traits\ObjectFitable;
-use Orchid\Attachment\Models\Attachment;
-use Orchid\Platform\Dashboard;
+use Czernika\OrchidImages\Screen\Components\Traits\CanBeEmpty;
+use Czernika\OrchidImages\Screen\Components\Traits\HasElements;
+use Czernika\OrchidImages\Screen\Components\Traits\ObjectFitable;
 use Orchid\Screen\Field;
 
+/**
+ * Undocumented class
+ */
 class Carousel extends Field
 {
-    use ObjectFitable;
+    use HasElements, ObjectFitable, CanBeEmpty;
 
     protected $view = 'orchid-images::components.carousel';
 
     protected $attributes = [
         'elements' => [],
+        'empty' => '',
         'controls' => false,
         'indicators' => false,
+        'lightbox' => false,
         'fade' => false,
-        'height' => 'auto',
-        'width' => '100%',
-        'empty' => '',
+        'height' => false,
+        'delay' => 5000,
         'fit' => 'object-fit-cover', // 'cover', 'contain', 'fill', 'scale', 'none'
     ];
 
     public function __construct()
     {
         $this->addBeforeRender(function () {
-            if (!empty($this->get('elements'))) {
-                return;
-            }
-
-            $value = $this->get('value');
-
-            $this->elements($value);
+            $this->setElements();
         });
     }
 
-    public function elements($elements)
-    {
-        if (is_numeric($elements)) {
-            $elements = Dashboard::model(Attachment::class)::find($elements);
-        }
-
-        $this->set('elements', Helper::isAttachment($elements) ? [$elements] : collect($elements));
-
-        return $this;
-    }
-
+    /**
+     * Set carousel control buttons
+     *
+     * @param boolean $controls
+     * @return static
+     */
     public function withControls(bool $controls = true): static
     {
-        $this->set('controls', $controls);
-
-        return $this;
+        return $this->set('controls', $controls);
     }
 
+    /**
+     * Set carousel bullets
+     *
+     * @param boolean $indicators
+     * @return static
+     */
     public function withIndicators(bool $indicators = true): static
     {
-        $this->set('indicators', $indicators);
-
-        return $this;
+        return $this->set('indicators', $indicators);
     }
 
+    /**
+     * Allow to use lightbox in carousel
+     *
+     * @param boolean $lightbox
+     * @return static
+     */
+    public function withLightbox(bool $lightbox = true): static
+    {
+        return $this->set('lightbox', $lightbox);
+    }
+
+    /**
+     * Set fade animation
+     *
+     * @param boolean $fade
+     * @return static
+     */
     public function fade(bool $fade = true): static
     {
-        $this->set('fade', $fade);
-
-        return $this;
+        return $this->set('fade', $fade);
     }
 
-    public function height($height): static
+    /**
+     * Set delay between slides animation
+     *
+     * @param integer $delay
+     * @return static
+     */
+    public function delay(int $delay): static
     {
-        $this->set('height', is_int($height) ? "{$height}px" : $height);
-
-        return $this;
+        return $this->set('delay', $delay);
     }
 
-    public function width($width): static
+    /**
+     * Set carousel height
+     *
+     * @param string|integer $height
+     * @return static
+     */
+    public function height(string|int $height): static
     {
-        $this->set('width', is_int($width) ? "{$width}px" : $width);
-
-        return $this;
-    }
-
-    public function empty(string $empty): static
-    {
-        $this->set('empty', $empty);
-
-        return $this;
+        return $this->set('height', is_numeric($height) ? "{$height}px" : $height);
     }
 }