+++ CHANGELOG.md
@@ -5,4 +5,4 @@
 - **refactor**: changed code base
 - **feat**: new options almost for every component
 - **feat**: added JS options for interactive elements such as Lightbox and Carousel
-- **tests**: added more cases
+- **tests**: covered more use cases